Fix/extend for-loops in Codie and fix parser templates
Now for-loop template has a valid `isLast` variable that says if current element is the last, obviously. Its behavor was wrong before, because of the block-scoping when exiting from inner loop, now it is fixed with wrapping for in an anonymous function. Also, I've change the end function of the template to actually return the value with what it wants to end the template, it is a bit more incapsulative, I think.